requirements.txt
langchain_openai
fastapi
python_dotenv
langchain_core
uvicorn